refactor: move dragging functionality into composable function
